:sparkles: Feat: log test 용 api 구현
<footer>
- 관련: #59
