+++ src/main/java/com/example/namo2/domain/test/ui/dto/TestResponse.java
@@ -1,0 +1,16 @@
+package com.example.namo2.domain.test.ui.dto;
+
+import lombok.Builder;
+
+public class TestResponse {
+
+	private TestResponse() {
+		throw new IllegalStateException("Utility class");
+	}
+
+	@Builder
+	public static class LogTestDto {
+		private String text;
+		private Integer number;
+	}
+}
